requirements.txt
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
mysql-connector==2.2.9
pycparser==2.20
PyJWT==2.0.1
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
